Combine DbDataReaderExtensions.GetColumnSchema implementations and remove unused facade
Closes #36235
