+++ src/System.Data.Common/src/System/Data/Common/DbDataReaderExtensions.cs
@@ -2,22 +2,85 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+
 namespace System.Data.Common
 {
     public static class DbDataReaderExtensions
     {
-        public static System.Collections.ObjectModel.ReadOnlyCollection<DbColumn> GetColumnSchema(this DbDataReader reader)
+        private sealed class DataRowDbColumn : DbColumn
         {
-            if (reader.CanGetColumnSchema())
+            private DataColumnCollection _schemaColumns;
+            private DataRow _schemaRow;
+
+            public DataRowDbColumn(DataRow readerSchemaRow, DataColumnCollection readerSchemaColumns)
             {
-                return ((IDbColumnSchemaGenerator)reader).GetColumnSchema();
+                _schemaRow = readerSchemaRow;
+                _schemaColumns = readerSchemaColumns;
+                PopulateFields();
             }
-            throw new NotSupportedException();
+
+            private void PopulateFields()
+            {
+                AllowDBNull = GetDbColumnValue<bool?>(SchemaTableColumn.AllowDBNull);
+                BaseCatalogName = GetDbColumnValue<string>(SchemaTableOptionalColumn.BaseCatalogName);
+                BaseColumnName = GetDbColumnValue<string>(SchemaTableColumn.BaseColumnName);
+                BaseSchemaName = GetDbColumnValue<string>(SchemaTableColumn.BaseSchemaName);
+                BaseServerName = GetDbColumnValue<string>(SchemaTableOptionalColumn.BaseServerName);
+                BaseTableName = GetDbColumnValue<string>(SchemaTableColumn.BaseTableName);
+                ColumnName = GetDbColumnValue<string>(SchemaTableColumn.ColumnName);
+                ColumnOrdinal = GetDbColumnValue<int?>(SchemaTableColumn.ColumnOrdinal);
+                ColumnSize = GetDbColumnValue<int?>(SchemaTableColumn.ColumnSize);
+                IsAliased = GetDbColumnValue<bool?>(SchemaTableColumn.IsAliased);
+                IsAutoIncrement = GetDbColumnValue<bool?>(SchemaTableOptionalColumn.IsAutoIncrement);
+                IsExpression = GetDbColumnValue<bool>(SchemaTableColumn.IsExpression);
+                IsHidden = GetDbColumnValue<bool?>(SchemaTableOptionalColumn.IsHidden);
+                IsIdentity = GetDbColumnValue<bool?>("IsIdentity");
+                IsKey = GetDbColumnValue<bool?>(SchemaTableColumn.IsKey);
+                IsLong = GetDbColumnValue<bool?>(SchemaTableColumn.IsLong);
+                IsReadOnly = GetDbColumnValue<bool?>(SchemaTableOptionalColumn.IsReadOnly);
+                IsUnique = GetDbColumnValue<bool?>(SchemaTableColumn.IsUnique);
+                NumericPrecision = GetDbColumnValue<int?>(SchemaTableColumn.NumericPrecision);
+                NumericScale = GetDbColumnValue<int?>(SchemaTableColumn.NumericScale);
+                UdtAssemblyQualifiedName = GetDbColumnValue<string>("UdtAssemblyQualifiedName");
+                DataType = GetDbColumnValue<Type>(SchemaTableColumn.DataType);
+                DataTypeName = GetDbColumnValue<string>("DataTypeName");
+            }
+
+            private T GetDbColumnValue<T>(string columnName) => _schemaColumns.Contains(columnName) && _schemaRow[columnName] is T value ? value : (default);
+        }
+
+        public static ReadOnlyCollection<DbColumn> GetColumnSchema(this DbDataReader reader)
+        {
+            if (reader is null)
+            {
+                throw new ArgumentNullException(nameof(reader));
+            }
+            if (reader is IDbColumnSchemaGenerator schemaGenerator)
+            {
+                return schemaGenerator.GetColumnSchema();
+            }
+            return GetColumnSchemaCompatibility(reader);
         }
 
         public static bool CanGetColumnSchema(this DbDataReader reader)
         {
-            return reader is IDbColumnSchemaGenerator;
+            return true;
+        }
+
+        private static ReadOnlyCollection<DbColumn> GetColumnSchemaCompatibility(DbDataReader reader)
+        {
+            var columnSchema = new List<DbColumn>();
+            DataTable schemaTable = reader.GetSchemaTable();
+            DataColumnCollection schemaTableColumns = schemaTable.Columns;
+            foreach (DataRow row in schemaTable.Rows)
+            {
+                columnSchema.Add(new DataRowDbColumn(row, schemaTableColumns));
+            }
+            return new ReadOnlyCollection<DbColumn>(columnSchema);
         }
     }
+
+
 }
